Inflate: Remove some AArch64-specific code
This is a ~0.3% regression on Apple M1 and a ~2% gain on Intel. Not a
huge deal either way, so the simplicity is nice to have.

This is possible by still using empty inline-assembly blocks to hint
the desired order of operations to the compiler, and because the
commit 'Inflate: rearrange shifts and ignore high bits of "bits"'
encourages the compiler to use 32-bit loads anyway.

Removing the "asm volatile" hints causes a 12% slowdown on M1.